lazyreload/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
auto-py-to-exe==2.35.0
awscli==1.18.69
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
botocore==1.16.19
bottle==0.12.25
bottle-websocket==0.2.9
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.2.0
Click==7.0
colorama==0.4.3
Columnar==1.4.1
command-not-found==0.3
contourpy==1.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.9
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
docutils==0.16
drf-yasg==1.21.7
duplicity==0.8.12.0
Eel==0.16.0
entrypoints==0.3
exceptiongroup==1.1.3
executing==1.2.0
Faker==23.2.1
fastapi==0.101.1
fasteners==0.14.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
gradio==3.40.1
gradio_client==0.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.14.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==2.8
importlib-resources==6.0.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.9.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linkify-it-py==2.0.2
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.5
multidict==6.0.4
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
openai==0.27.8
orjson==3.9.5
packaging==23.1
pandas==2.0.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==3.6.1
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.4.2
pycairo==1.16.2
pycups==1.9.73
pydantic==2.2.1
pydantic_core==2.6.1
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.36.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
PyPDF2==3.0.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.2
pyqt6-tools==6.4.2.3.3
pyRFC3339==1.1
pytest==7.4.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
referencing==0.30.2
regex==2023.12.25
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
roman==2.0.0
rpds-py==0.9.2
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
semantic-version==2.10.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
string-color==1.2.3
systemd-python==234
tiktoken==0.4.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.23.2
wadllib==1.3.3
wcwidth==0.2.6
websockets==11.0.3
whichcraft==0.6.1
xkit==0.0.0
yarl==1.9.2
zipp==3.16.2
zope.event==4.6
zope.interface==6.0
